Allow binding to a specific interface
statsd currently binds to all (IPv4) interfaces.  This may not always be what we want
(it's not what I want).  Sometimes we just want to bind to a private interface so that
only boxes on our local network can send us messages.

This patch lets us do that.

Note that we default to undefined.  Both dgram.js and net.js treat undefined as the
default value, ie, bind to everything.