Add gauge types.
Some applications have absolute values that they'd like to track as a simple
gauge, rather than as a count per period. This change adds support for this
type. Clients can set a gauge by sending 'value|g'.

Any gauge value which is set will persist and be flushed out to graphite each
iteration. This means that the gauge can be set once, and that value will be
graphed continuously until the next time it's set.